씬 컨트롤러 초기화 수정

+++ Assets/Seongho/Scripts/Controller/CUIPlayerController.cs
@@ -24,7 +24,7 @@ public class CUIPlayerController : CPlayerController
 
     public override int GetHorizontal()
     {
-        int value = (int)mJoyStick.value;
+        float value = mJoyStick.value;
 
         if (value < 2)
         {
